pyproject.toml
[project]
name = "ai-llm-adversarial-conversation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.68.0",
    "beautifulsoup4>=4.13.5",
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "gradio>=5.47.0",
    "ipython>=9.5.0",
    "ollama>=0.5.4",
    "openai>=1.107.1",
    "requests>=2.32.5",
]
